backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
mysql-connector-python
python-dotenv==1.0.0
python-dateutil==2.8.2
bcrypt==4.1.2
